feat(Layout): load user data from API and pass to UserCard component
